web/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
asn1crypto==0.24.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.3
beautifulsoup4==4.4.0
billiard==3.3.0.22
boto==2.40.0
ccm==2.1.4
celery==3.1.19
certifi==2016.2.28
cffi==1.9.1
characteristic==14.3.0
cryptography==2.1.4
cssselect==0.9.1
cycler==0.10.0
da-vinci==0.2.2
dask==0.8.2
decorator==4.0.9
diff-match-patch==20121119
Django==1.9
django-activity-stream==0.6.0
django-allauth==0.24.1
django-appconf==1.0.2
django-celery==3.1.17
django-chartit==0.2.9
django-countries==5.0
django-filter==1.1.0
django-grappelli==2.6.3
django-import-export==0.4.5
django-ipware==0.1.0
django-mailgun==0.2.2
django-model-utils==2.4
django-notifications-hq==1.0
django-paypal==0.2
django-profiles==0.2
django-queued-storage==0.8.0
django-recaptcha==1.0.3
django-registration==1.0
django-registration-redux==1.1
django-responsive==0.3.0
django-simple-captcha==0.5.5
django-social-friends-finder==0.2
django-socialauth==0.1.2rc0
django-storages==1.4.1
django-taggit==0.18.1
django-taggit-serializer==0.1.5
django-thumbnails==0.2.0
django-tinymce==2.3.0
django-user-agents==0.3.2
django-widget-tweaks==1.4.1
django-wysiwyg==0.7.1
djangorestframework==3.7.0
djorm-ext-pgfulltext==0.10
easy-thumbnails==2.4.1
ecdsa==0.13
enum34==1.1.6
et-xmlfile==1.0.1
Fabric==1.14.0
flup==1.0.2
functools32==3.2.3.post2
future==0.16.0
gunicorn==19.7.1
hashids==1.2.0
httplib2==0.10.3
idna==2.2
ipaddress==1.0.18
jsonfield==1.0.3
kombu==3.0.33
mysqlclient==1.3.12
oauthlib==1.0.3
paramiko==2.3.1
Pillow==2.7.0
pika==0.11.2
pyasn1==0.1.8
pycountry==17.9.23
pycparser==2.17
PyNaCl==1.1.2
python-dateutil==2.6.1
python-memcached==1.58
python-openid==2.2.5
pytz==2017.2
PyYAML==3.11
rcssmin==1.0.6
requests==2.10.0
requests-oauthlib==0.6.0
restclient==0.11.0
shortuuid==0.5.0
six==1.10.0
sorl-thumbnail==12.4.1
sorl-thumbnail-serializer-field==0.2.1
South==1.0.2
tablib==0.11.2
ua-parser==0.7.3
user-agents==1.1.0
vine==1.1.4
whois==0.7
django-geoposition